# packages/shared-types/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/intent": {
      "post": {
        "operationId": "IntentController_dispatch",
        "summary": "Dispatch intent to AI service",
        "description": "Returns request_id for SSE stream pickup via GET /intent/stream?id=<rid>. S-04 T03: mode field selects Variant B (ai_augmented) or Variant A (basic_fallback) per D-S04-03 LAW Adaptive Single Endpoint. Sx05-3-CODE HOTFIX (D-S05-13 LAW): @UseGuards(JwtAuthGuard) added — req.user.id forwarded to AI as PostIntentBody.user_id so cart_by_text graph operates on correct authenticated cart.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IntentRequestDto"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": ""
          }
        },
        "tags": [
          "intent"
        ],
        "security": [
          {
            "icp_session": []
          }
        ]
      }
    },
    "/api/v1/intent/stream": {
      "get": {
        "operationId": "IntentController_stream",
        "summary": "SSE stream of intent events",
        "description": "Forwards Redis pub/sub channel `sse:pubsub:{rid}` to FE EventSource. S-04 T03: 17 typed event types (10 S-02 + 7 S-04) per 03_API §3. Requires icp_session cookie.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "intent"
        ],
        "security": [
          {
            "icp_session": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "ICP Gateway API",
    "description": "Intelligent Commerce Platform — REST API.\n\n4 mutating endpoints require `Idempotency-Key` header (UUID v4) per docs/03_API_CONTRACTS.md §1.",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Local dev"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "icp_session",
        "description": "httpOnly session cookie issued by POST /auth/login (per ADR-019 + S-03 C-01)."
      }
    },
    "schemas": {
...
      "IntentRequestDto": {
        "type": "object",
        "properties": {
          "modality": {
            "type": "string",
            "enum": [
              "text",
              "image",
              "voice"
            ]
          },
          "content": {
            "type": "string",
            "minLength": 1,
            "maxLength": 10000000
          },
          "hint": {
            "type": "string",
            "enum": [
              "import",
              "buy",
              "search",
              "recommend",
              "cart_clear_confirm",
              "cart_view_with_stock_check",
              "analyze"
            ]
          },
          "mode": {
            "type": "string",
            "enum": [
              "ai_augmented",
              "basic_fallback"
            ],
            "default": "ai_augmented"
          },
          "text_query": {
            "type": "string",
            "minLength": 1,
            "maxLength": 500
          },
          "filters": {
            "type": "object",
            "properties": {
              "brand": {
                "type": "string",
                "minLength": 1,
                "maxLength": 100
              },
              "category": {
                "type": "string",
                "minLength": 1,
                "maxLength": 50
              },
              "attributes": {
                "type": "object",
                "additionalProperties": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 100
                }
              }
            }
          }
        },
        "required": [
          "modality"
        ]
      },
...
    }
  }
}